Avoid returning nulls when the result is not found
